Refactor api tests and rename config to connection
